refactor: move stake button component to stake directory
