+++ app/components/UI/Stake/components/StakeButton/index.tsx
@@ -1,14 +1,11 @@
 import React from 'react';
-import { TokenI, BrowserTab } from '../../types';
+import { TokenI, BrowserTab } from '../../../Tokens/types';
 import { useNavigation } from '@react-navigation/native';
-import { isPooledStakingFeatureEnabled } from '../../../Stake/constants';
+import { isPooledStakingFeatureEnabled } from '../../constants';
 import Routes from '../../../../../constants/navigation/Routes';
 import { useSelector } from 'react-redux';
 import AppConstants from '../../../../../core/AppConstants';
-import {
-  MetaMetricsEvents,
-  useMetrics,
-} from '../../../../../components/hooks/useMetrics';
+import { MetaMetricsEvents, useMetrics } from '../../../../hooks/useMetrics';
 import { getDecimalChainId } from '../../../../../util/networks';
 import { selectChainId } from '../../../../../selectors/networkController';
 import { Pressable } from 'react-native';
@@ -18,7 +15,7 @@ import Text, {
 } from '../../../../../component-library/components/Texts/Text';
 import { WalletViewSelectorsIDs } from '../../../../../../e2e/selectors/wallet/WalletView.selectors';
 import { useTheme } from '../../../../../util/theme';
-import createStyles from '../../styles';
+import createStyles from '../../../Tokens/styles';
 import Icon, {
   IconColor,
   IconName,
@@ -26,8 +23,8 @@ import Icon, {
 } from '../../../../../component-library/components/Icons/Icon';
 import { strings } from '../../../../../../locales/i18n';
 import { RootState } from '../../../../../reducers';
-import useStakingEligibility from '../../../Stake/hooks/useStakingEligibility';
-import { StakeSDKProvider } from '../../../Stake/sdk/stakeSdkProvider';
+import useStakingEligibility from '../../hooks/useStakingEligibility';
+import { StakeSDKProvider } from '../../sdk/stakeSdkProvider';
 
 interface StakeButtonProps {
   asset: TokenI;
